+++ crates/proto/src/domain/account.rs
@@ -3,6 +3,7 @@ use std::fmt::{Debug, Display, Formatter};
 use miden_node_utils::formatting::format_opt;
 use miden_objects::Word;
 use miden_objects::account::{Account, AccountHeader, AccountId};
+use miden_objects::asset::Asset;
 use miden_objects::block::{AccountWitness, BlockNumber};
 use miden_objects::note::{NoteExecutionMode, NoteTag};
 use miden_objects::utils::{Deserializable, DeserializationError, Serializable};
@@ -280,6 +281,28 @@ impl TryFrom<proto::block_producer_store::transaction_inputs::AccountTransaction
         };
 
         Ok(Self { account_id, account_commitment })
+    }
+}
+
+// ASSET
+// ================================================================================================
+
+impl From<Asset> for proto::primitives::Asset {
+    fn from(asset_from: Asset) -> Self {
+        proto::primitives::Asset {
+            asset: Some(Word::from(asset_from).into()),
+        }
+    }
+}
+
+impl TryFrom<proto::primitives::Asset> for Asset {
+    type Error = ConversionError;
+
+    fn try_from(value: proto::primitives::Asset) -> Result<Self, Self::Error> {
+        let inner = value.asset.ok_or(proto::primitives::Asset::missing_field("asset"))?;
+        let word = Word::try_from(inner)?;
+
+        Asset::try_from(word).map_err(ConversionError::AssetError)
     }
 }