+++ crates/proto/src/errors.rs
@@ -1,15 +1,17 @@
 use std::any::type_name;
 use std::num::TryFromIntError;
 
-use miden_objects::FeeError;
 use miden_objects::crypto::merkle::{SmtLeafError, SmtProofError};
 use miden_objects::utils::DeserializationError;
+use miden_objects::{AssetError, FeeError};
 use thiserror::Error;
 
 use crate::domain::note::NetworkNoteError;
 
 #[derive(Debug, Error)]
 pub enum ConversionError {
+    #[error("asset error")]
+    AssetError(#[from] AssetError),
     #[error("fee parameters error")]
     FeeError(#[from] FeeError),
     #[error("hex error")]