+++ crates/store/src/db/models/queries/accounts.rs
@@ -18,13 +18,13 @@ use miden_node_proto as proto;
 use miden_node_proto::domain::account::{AccountInfo, AccountSummary};
 use miden_node_utils::limiter::{QueryParamAccountIdLimit, QueryParamLimiter};
 use miden_objects::account::{Account, AccountCode, AccountId, AccountStorage};
-use miden_objects::asset::AssetVault;
+use miden_objects::asset::{Asset, AssetVault};
 use miden_objects::block::BlockNumber;
 use miden_objects::{Felt, Word};
 
 use crate::db::models::conv::{SqlTypeConvert, raw_sql_to_nonce, raw_sql_to_slot};
 use crate::db::models::{serialize_vec, vec_raw_try_into};
-use crate::db::schema;
+use crate::db::{AccountVaultValue, schema};
 use crate::errors::DatabaseError;
 
 /// Select the latest account details by account id from the DB using the given
@@ -149,6 +149,81 @@ pub(crate) fn select_accounts_by_id(
     Ok(account_infos)
 }
 
+pub(crate) fn select_account_vault_assets(
+    conn: &mut SqliteConnection,
+    account_id: AccountId,
+    block_from: BlockNumber,
+    block_to: BlockNumber,
+) -> Result<(BlockNumber, Vec<AccountVaultValue>), DatabaseError> {
+    use schema::account_vault_assets as t;
+
+    // SELECT
+    //     block_num,
+    //     vault_key,
+    //     asset
+    // FROM
+    //     account_vault_assets
+    // WHERE
+    //     account_id = ?
+    //     AND block_num >= ?
+    //     AND block_num <= ?
+    // ORDER BY
+    //     block_num ASC
+    // LIMIT
+    //     ROW_LIMIT;
+
+    // TODO: These limits should be given by the protocol.
+    // See miden-base/issues/1770 for more details
+    const MAX_PAYLOAD_BYTES: usize = 2 * 1024 * 1024; // 2 MB
+    const ROW_OVERHEAD_BYTES: usize = 2 * size_of::<Word>() + size_of::<u32>(); // key + asset + block_num
+    const ROW_LIMIT: usize = (MAX_PAYLOAD_BYTES / ROW_OVERHEAD_BYTES) + 1;
+
+    if !account_id.is_public() {
+        return Err(DatabaseError::AccountNotPublic(account_id));
+    }
+
+    if block_from > block_to {
+        return Err(DatabaseError::InvalidBlockRange { from: block_from, to: block_to });
+    }
+
+    let raw: Vec<(i64, Vec<u8>, Option<Vec<u8>>)> =
+        SelectDsl::select(t::table, (t::block_num, t::vault_key, t::asset))
+            .filter(
+                t::account_id
+                    .eq(account_id.to_bytes())
+                    .and(t::block_num.ge(block_from.to_raw_sql()))
+                    .and(t::block_num.le(block_to.to_raw_sql())),
+            )
+            .order(t::block_num.asc())
+            .limit(i64::try_from(ROW_LIMIT).expect("should fit within i64"))
+            .load::<(i64, Vec<u8>, Option<Vec<u8>>)>(conn)?;
+
+    // Discard the last block in the response (assumes more than one block may be present)
+    let (last_block_included, values) = if raw.len() >= ROW_LIMIT {
+        // NOTE: If the query contains at least one more row than the amount of storage map updates
+        // allowed in a single block for an account, then the response is guaranteed to have at
+        // least two blocks
+
+        // SAFETY: we checked that the vector is not empty
+        let &(last_block_num, ..) = raw.last().unwrap();
+
+        let values = raw
+            .into_iter()
+            .take_while(|(bn, ..)| *bn != last_block_num)
+            .map(AccountVaultValue::from_raw_row)
+            .collect::<Result<Vec<_>, DatabaseError>>()?;
+
+        (BlockNumber::from_raw_sql(last_block_num.saturating_sub(1))?, values)
+    } else {
+        (
+            block_to,
+            raw.into_iter().map(AccountVaultValue::from_raw_row).collect::<Result<_, _>>()?,
+        )
+    };
+
+    Ok((last_block_included, values))
+}
+
 /// Select [`AccountSummary`] from the DB using the given [`SqliteConnection`], given that the
 /// account update was done between `(block_start, block_end]`.
 ///
@@ -175,6 +250,7 @@ pub fn select_accounts_by_block_range(
     //     account_id IN rarray(?3)
     // ORDER BY
     //     block_num ASC
+
     let desired_account_ids = serialize_vec(account_ids);
     let raw: Vec<AccountSummaryRaw> =
         SelectDsl::select(schema::accounts::table, AccountSummaryRaw::as_select())
@@ -285,7 +361,8 @@ pub(crate) fn select_account_storage_map_values(
     // TODO: These limits should be given by the protocol.
     // See miden-base/issues/1770 for more details
     pub const MAX_PAYLOAD_BYTES: usize = 2 * 1024 * 1024; // 2 MB
-    pub const ROW_OVERHEAD_BYTES: usize = size_of::<Word>() + size_of::<Word>() + size_of::<u8>(); // key + value + slot_idx
+    pub const ROW_OVERHEAD_BYTES: usize =
+        2 * size_of::<Word>() + size_of::<u32>() + size_of::<u8>(); // key + value + block_num + slot_idx
     pub const ROW_LIMIT: usize = (MAX_PAYLOAD_BYTES / ROW_OVERHEAD_BYTES) + 1;
 
     if !account_id.is_public() {
@@ -309,6 +386,7 @@ pub(crate) fn select_account_storage_map_values(
             .load(conn)?;
 
     // Discard the last block in the response (assumes more than one block may be present)
+
     let (last_block_included, values) = if raw.len() >= ROW_LIMIT {
         // NOTE: If the query contains at least one more row than the amount of storage map updates
         // allowed in a single block for an account, then the response is guaranteed to have at
@@ -331,6 +409,27 @@ pub(crate) fn select_account_storage_map_values(
     };
 
     Ok(StorageMapValuesPage { last_block_included, values })
+}
+
+#[derive(Queryable, Selectable)]
+#[diesel(table_name = crate::db::schema::account_vault_assets)]
+#[diesel(check_for_backend(diesel::sqlite::Sqlite))]
+pub struct AccountVaultUpdateRaw {
+    pub vault_key: Vec<u8>,
+    pub asset: Option<Vec<u8>>,
+    pub block_num: i64,
+}
+
+impl TryFrom<AccountVaultUpdateRaw> for AccountVaultValue {
+    type Error = DatabaseError;
+
+    fn try_from(raw: AccountVaultUpdateRaw) -> Result<Self, Self::Error> {
+        let vault_key = Word::read_from_bytes(&raw.vault_key)?;
+        let asset = raw.asset.map(|bytes| Asset::read_from_bytes(&bytes)).transpose()?;
+        let block_num = BlockNumber::from_raw_sql(raw.block_num)?;
+
+        Ok(AccountVaultValue { block_num, vault_key, asset })
+    }
 }
 
 #[derive(Debug, Clone, Queryable, QueryableByName, Selectable)]