+++ crates/store/src/errors.rs
@@ -14,6 +14,8 @@ use miden_objects::{
     AccountDeltaError,
     AccountError,
     AccountTreeError,
+    AssetError,
+    AssetVaultError,
     FeeError,
     NoteError,
     NullifierTreeError,
@@ -39,6 +41,10 @@ pub enum DatabaseError {
     AccountError(#[from] AccountError),
     #[error("account delta error")]
     AccountDeltaError(#[from] AccountDeltaError),
+    #[error("asset vault error")]
+    AssetVaultError(#[from] AssetVaultError),
+    #[error("asset error")]
+    AssetError(#[from] AssetError),
     #[error("closed channel")]
     ClosedChannel(#[from] RecvError),
     #[error("deserialization failed")]