+++ crates/utils/src/tracing/grpc.rs
@@ -47,6 +47,7 @@ fn rpc_trace_fn<T>(request: &http::Request<T>) -> tracing::Span {
         Some("CheckNullifiersByPrefix") => rpc_span!("rpc.rpc", "CheckNullifiersByPrefix"),
         Some("GetBlockHeaderByNumber") => rpc_span!("rpc.rpc", "GetBlockHeaderByNumber"),
         Some("SyncStorageMaps") => rpc_span!("rpc.rpc", "SyncStorageMaps"),
+        Some("SyncAccountVault") => rpc_span!("rpc.rpc", "SyncAccountVault"),
         Some("SyncState") => rpc_span!("rpc.rpc", "SyncState"),
         Some("SyncNotes") => rpc_span!("rpc.rpc", "SyncNotes"),
         Some("GetNotesById") => rpc_span!("rpc.rpc", "GetNotesById"),